AI_project/scholarship-image-recognition/requirements.txt
python-gemini-api

